Fix simulation artifact by change variable format
Variable format of kx and ky changed back to float (previously double)
to fix the artifact. Artifact fixed when doing 20*20 simulation with no
aberration, but the simulation used chi0 = -chi(), this version changed
back to chi0 = chi() to follow the routine of old autostem.c

+++ autostem.hpp
@@ -222,7 +222,7 @@ private:
         cfpix *probe, trans;     // complex probes and transmission functions
 
         float zmin, zmax;
-        double *kx, *ky, *kx2, *ky2, *kxp, *kyp, *kxp2, *kyp2;
+        float *kx, *ky, *kx2, *ky2, *kxp, *kyp, *kxp2, *kyp2;
         float *propxr, *propxi;   /* complex propagator vs x */
         float *propyr, *propyi;   /* complex propagator vs y */
         float *xp, *yp;
